+++ src/graphql.tsx
@@ -641,7 +641,7 @@ class ObservableQueryRecycler {
     // Stop the query from polling when we recycle. Polling may resume when we
     // reuse it and call `setOptions`.
     observableQuery.setOptions({
-      fetchPolicy: 'cache-only',
+      fetchPolicy: 'standby',
       pollInterval: 0,
     });